+++ src/trtutils/impls/yolo/_yolo.py
@@ -27,7 +27,6 @@ class YOLO:
         warmup_iterations: int = 10,
         *,
         warmup: bool | None = None,
-        scale_inputs: bool | None = None,
     ) -> None:
         """
         Create a YOLO object.
@@ -44,9 +43,6 @@ class YOLO:
             The default is 10.
         warmup : bool, optional
             Whether or not to perform warmup iterations.
-        scale_inputs : bool, optional
-            Whether or not to scale inputs to [0.0:1.0] range.
-            By default will perform scaling.
 
         Raises
         ------
@@ -75,7 +71,6 @@ class YOLO:
             raise ValueError(err_msg)
         self._input_size: tuple[int, int] = (input_size[3], input_size[2])
         self._dtype = input_spec[1]
-        self._scale_inputs = scale_inputs
 
     @property
     def engine(self: Self) -> TRTEngine:
@@ -117,7 +112,9 @@ class YOLO:
         """
         _log.debug(f"{self._tag}: Running preprocess")
         return preprocess(
-            image, self._input_size, self._dtype, scale_down=self._scale_inputs,
+            image,
+            self._input_size,
+            self._dtype,
         )
 
     def postprocess(